Split app.ts into multiple files
- app.ts handles events calls necessary functions. Always returns valid
http response.
- src/*/ contains each event type handlers
- src/utils/ contains helper functions
