+++ reward-monitor-lambda-app/app.ts
@@ -1,10 +1,9 @@
-import { APIGatewayEvent, APIGatewayProxyResult, Context, EventBridgeEvent } from 'aws-lambda';
+import { APIGatewayEvent, APIGatewayProxyCallback, APIGatewayProxyResult, Context, EventBridgeEvent } from 'aws-lambda';
 
 import { DynamoDBClient } from '@aws-sdk/client-dynamodb';
 import { DynamoDBDocumentClient, PutCommand } from '@aws-sdk/lib-dynamodb';
-
-import { ethers } from 'ethers';
-import { getMinterOwedPutItem } from './src/minterDebtHelper';
+import { runDailyJob } from './src/cloudwatch/daily';
+import { handleCheck } from './src/api/check';
 
 const TABLE_NAME = process.env.TABLE_NAME as string;
 
@@ -21,41 +20,52 @@ export const lambdaHandler = async (
     event: APIGatewayEvent | EventBridgeEvent<string, any>,
     context: Context,
 ): Promise<APIGatewayProxyResult> => {
-    if ('source' in event) {
-        console.log('EventBridge event received');
-        console.log(JSON.stringify(event, null, 2));
-    } else {
-        console.log('API Gateway event received');
-        console.log(JSON.stringify(event, null, 2));
-    }
     // Create client objects
     const ddbClient = new DynamoDBClient({});
     const ddbDocClient = DynamoDBDocumentClient.from(ddbClient);
 
-    // Create an ethers provider
-    const mainnetProvider = new ethers.providers.JsonRpcProvider(
-        `https://eth-mainnet.alchemyapi.io/v2/${process.env.ALCHEMY_API_KEY}`,
-    );
-
-    // Build the params object for the DynamoDB PutItem command
-    console.log('Getting minter owed data...');
-
-    // Get the minter owed data as a DynamoDB PutItem params object
-    const params = await getMinterOwedPutItem(mainnetProvider, TABLE_NAME).catch((err) => {
-        console.error(`Failed to get minter owed data. Error: ${err}`);
-        throw err;
-    });
-
-    const response = {
+    let response = {
         statusCode: 200,
-        body: JSON.stringify(params.Item),
+        body: "No job was run. This is likely because the API path was not '/check'",
     };
 
-    // Save the data to DynamoDB
-    await ddbDocClient.send(new PutCommand(params)).catch((err) => {
-        response.statusCode = 500;
-        response.body = `Failed to save data to DynamoDB. Error: ${err}`;
-    });
+    // Check if the event is from EventBridge (CloudWatch) or API Gateway
+    if ('source' in event) {
+        // EventBridge (CloudWatch) event
+        console.log('EventBridge event received');
+
+        // Run the daily job and return the response
+        // If the job fails, the error will be caught and logged
+        response = await runDailyJob(ddbDocClient).catch((err) => {
+            const errorMessage = `Failed to run daily job. Error: ${JSON.stringify(err)}`;
+            console.error(errorMessage);
+            return {
+                statusCode: 500,
+                body: errorMessage,
+            };
+        });
+    } else {
+        // API Gateway event
+        console.log('API Gateway event received');
+
+        // Check if the path is /check and handle it
+        if (event.path.toLowerCase() === '/check') {
+            console.log('Handling /check');
+            response = await handleCheck(ddbDocClient).catch((err) => {
+                const errorMessage = `Failed to handle /check. Error: ${JSON.stringify(err)}`;
+                console.error(errorMessage);
+                return {
+                    statusCode: 500,
+                    body: errorMessage,
+                };
+            });
+        } else {
+            response = {
+                statusCode: 404,
+                body: 'Path not found',
+            };
+        }
+    }
 
     return response;
 };
