+++ reward-monitor-lambda-app/src/api/check.ts
@@ -1,0 +1,5 @@
+import { APIGatewayProxyResult } from 'aws-lambda';
+
+export const handleCheck = async (ddbDocClient: DynamoDBDocumentClient): Promise<APIGatewayProxyResult> => {
+    
+};
